+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0] - 2024-04-07
+
+### Added
+
+- [Data] - Level 7 secret words;
+
+### Fixed
+
+- [Data] - Secret word fix with 15 clues;
+
 ## [1.0.0] - 2024-03-31
 
 ### Added